requirements.txt
fastmcp==0.2.0
yfinance==0.2.28
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.2